runtime code refactor
